TM-87: Use KnowledgeErrors
